# #2372 - Restrict Redis Checks