--- a/.talismanrc
+++ b/.talismanrc
@@ -17,12 +17,14 @@ fileignoreconfig:
   checksum: 4e6e2fabf9630e9e0d26028818f41f34c065e80c223db34d35a1d1068c37a782
 - filename: poetry.lock
   checksum: 5369b357fbab76c1aca03f779132c86d3894fcf8c1e2dbe26ba4de30a418f531
+- filename: app/notifications/validators.py
+  checksum: 2a926e5fa79eadccae3cd4dc423c7f83b0870d83719bb5d6d1f32b0793e5af31
 - filename: tests/app/authentication/test_authentication.py
   checksum: f7091cd49ae042cb1a189d81e905c7595a822e5a6742624ed37a0ed7e1e05bb5
 - filename: tests/app/delivery/test_send_to_providers.py
   checksum: 7db1b20912c3f7a5cb366eb143cace80bfb406ca1af844e11022747e7311aea5
 - filename: tests/app/notifications/test_validators.py
-  checksum: 7e95b0854fdc12ab60e97347f5e6498e99b29b9df3b635487ab89189ed11fabb
+  checksum: 9dab7e3890137b515bf84f9ea587d79dbd5e63568ff9a72c3a404f7738c0b329
 - filename: tests/app/service/test_rest.py
   checksum: b2126023c505e04dc66e4cc2be876c49daad154acd0c2ac63f03961e2ea8bef2
 - filename: tests/lambda_functions/nightly_stats_bigquery_upload/test_nightly_lambda.py