--- a/.talismanrc
+++ b/.talismanrc
@@ -18,7 +18,7 @@ fileignoreconfig:
 - filename: poetry.lock
   checksum: 5369b357fbab76c1aca03f779132c86d3894fcf8c1e2dbe26ba4de30a418f531
 - filename: app/notifications/validators.py
-  checksum: 2a926e5fa79eadccae3cd4dc423c7f83b0870d83719bb5d6d1f32b0793e5af31
+  checksum: 7c3f9574dd8096863b53927723fccc75f8c0a27c06a2da09b277fbc8af7ba9ef
 - filename: tests/app/authentication/test_authentication.py
   checksum: f7091cd49ae042cb1a189d81e905c7595a822e5a6742624ed37a0ed7e1e05bb5
 - filename: tests/app/delivery/test_send_to_providers.py