<!--
Before you open this PR, make sure that you review and are taking steps to follow [the Definition of Mergeable in our team agreement](https://docs.google.com/document/d/1nwZIF_lydPWfvixxZlQLNt4nqy3Qp13pHQnMcYJjTqE/edit#heading=h.6mnhaqm79e12). You may need to scroll down to that subheader.
-->

# Description
<!--
Please include a summary of the change and which issue is fixed. Please also include relevant motivation and context. List any dependencies that are required for this change.
-->

This PR moved the redis daily limit increments to the daily message limit checking code, rather than the persist notification code. Tests that checked the redis increments when persisting notifications have been removed and tests for the daily limit checks have been simplified.

issue #2372 

## How Has This Been Tested?
<!--
Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also list any relevant details for your test configuration
How has this been tested? (e.g. Unit tests, Tested locally, Tested as a GitHub action, Depoyed to dev, etc)  
Please provide relevant information and / or links to demonstrate the functionality or fix. (e.g. screenshots, link to deployment, regression test results, etc)
-->

- [x] deployed to dev
- [x] test logs
	- [x] approaching daily limit > 75% log
	- [x] daily limit exceeded log
- [x] ensure daily limits still work as intended

We can see that rate limit is working here. Set the limit of the Test service to 5 temporarily.
![image](https://github.com/user-attachments/assets/62b4bbdf-84ac-4bf9-bd1b-f45b6a4b8505)

### Test Timing Improvements
There were some significant improvements in some tests that were attempting to connect to redis.

**tests/app/notifications/test_process_notifications.py**
**before:**
```
=============== slowest 5 durations ===============
4.20s call     tests/app/notifications/test_process_notifications.py::test_persist_notification_creates_and_save_to_db
3.29s call     tests/app/notifications/test_process_notifications.py::test_persist_notification_with_optionals
1.20s call     tests/app/notifications/test_process_notifications.py::test_send_notification_to_queue_delayed_nonzero_delay
0.36s call     tests/app/notifications/test_process_notifications.py::test_send_notification_to_queue_delayed_throws_exception_when_missing_queue
0.34s call     tests/app/notifications/test_process_notifications.py::test_persist_notification_persists_recipient_identifiers[VAPROFILEID-some va profile id-email]
=============== 85 passed in 27.93s ===============
```

**after:**
```
=============== slowest 5 durations ===============
1.20s call     tests/app/notifications/test_process_notifications.py::test_send_notification_to_queue_delayed_nonzero_delay
0.35s setup    tests/app/notifications/test_process_notifications.py::test_send_notification_to_queue_delayed_zero_delay
0.32s call     tests/app/notifications/test_process_notifications.py::test_persist_notification_creates_and_save_to_db
0.32s call     tests/app/notifications/test_process_notifications.py::test_persist_notification_persists_recipient_identifiers[ICN-some icn-sms]
0.31s call     tests/app/notifications/test_process_notifications.py::test_send_notification_to_queue_throws_exception_deletes_notification
=============== 81 passed in 19.77s ===============
```


## Checklist

- [x] I have assigned myself to this PR
- [x] PR has an [appropriate title](https://github.com/department-of-veterans-affairs/vanotify-team/blob/main/Engineering/team_agreement.md#naming-prs): #9999 - What the thing does
- [x] PR has a detailed description, including links to specific documentation
- [x] I have added the [appropriate labels](https://github.com/department-of-veterans-affairs/vanotify-team/blob/main/Engineering/pull_request_labels.md#vanotify-labels) to the PR.
- [x] I did not remove any parts of the template, such as checkboxes even if they are not used
- [x] My code follows the [style guidelines](https://github.com/department-of-veterans-affairs/vanotify-team/blob/main/Process/style_guide.md) of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to any documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works. [Testing guidelines](https://github.com/department-of-veterans-affairs/vanotify-team/blob/main/Process/testing_guide.md)
- [x] I have ensured the latest main is merged into my branch and all checks are green prior to review
- [x] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules
- [x] The ticket was moved into the DEV test column when I began testing this change